-refactor code for the comment section -delete comment only for the comment post by the logged user -refactor service to check user do legal action action are done

+++ src/main/java/api/service/REST/CommentRESTController.java
@@ -1,6 +1,7 @@
 package api.service.REST;
 
 import api.dto.CommentDTO;
+import api.dto.UserDTO;
 import core.service.CommentService;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,6 +21,9 @@ public class CommentRESTController {
     private CommentService commentService;
 
     private static final Logger logger = Logger.getLogger(CommentRESTController.class);
+
+    @Autowired
+    private UserDTO userDTO;
 
     @RequestMapping(value = "/{videoId}", method = RequestMethod.GET, produces = "application/json", consumes = "application/json")
     public @ResponseBody List<CommentDTO> getCommentVideoPaginated(@PathVariable("videoId")int videoId,
@@ -41,28 +45,64 @@ public class CommentRESTController {
     public ResponseEntity createComment(@RequestBody CommentDTO comment) {
 
         try{
-            commentService.createComment(comment);
+
+            if(userDTO != null && userDTO.getId().equals(comment.getUserId())) {
+                commentService.createComment(comment);
+            }
+            else if(userDTO != null) {
+                logger.error(String.format("user try to insert comment for other people userId trying [%s] " +
+                        "comment userId [%s]", userDTO.getId(), comment.getUserId()));
+
+                return new ResponseEntity<String>("Erreur", new HttpHeaders(), HttpStatus.UNAUTHORIZED);
+            }
+            else {
+                logger.error(String.format("user unlogged to post comment for other people" +
+                        "comment userId [%s]", comment.getUserId()));
+
+                return new ResponseEntity<String>("Erreur", new HttpHeaders(), HttpStatus.UNAUTHORIZED);
+            }
         }catch (Exception e){
             logger.error("Error while creation comment" + e);
-            return new ResponseEntity("Erreur", new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<String>("Erreur", new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
 
 
-        return new ResponseEntity("OK", new HttpHeaders(), HttpStatus.CREATED);
+        return new ResponseEntity<String>("OK", new HttpHeaders(), HttpStatus.CREATED);
     }
 
     @RequestMapping(value="/{id}", method = RequestMethod.DELETE, produces = "application/json", consumes = "application/json")
     public ResponseEntity deleteComment(@PathVariable("id")int commentID) {
 
         try{
-            commentService.deleteComment(commentID);
+            CommentDTO commentDTO = commentService.findById(commentID);
+            if(commentDTO != null){
+                if(userDTO != null && userDTO.getId().equals(commentDTO.getUserId())) {
+                    commentService.deleteComment(commentID);
+                }
+                else if(userDTO != null) {
+                    logger.error(String.format("user try to delete other people comment  userId trying [%s] " +
+                                                "comment userId [%s]", userDTO.getId(), commentDTO.getUserId()));
+
+                    return new ResponseEntity<String>("Erreur", new HttpHeaders(), HttpStatus.UNAUTHORIZED);
+                }
+                else {
+                    logger.error(String.format("user unlogged to delete other people comment" +
+                            "comment userId [%s]", commentDTO.getUserId()));
+
+                    return new ResponseEntity<String>("Erreur", new HttpHeaders(), HttpStatus.UNAUTHORIZED);
+                }
+            }
+            else {
+                logger.error(String.format("comment not found for id [%s]", commentID));
+                return new ResponseEntity<String>("Erreur", new HttpHeaders(), HttpStatus.NOT_FOUND);
+            }
         }catch (Exception e){
             logger.error("Error while deleting comment" + e);
-            return new ResponseEntity("Erreur", new HttpHeaders(), HttpStatus.NOT_FOUND);
+            return new ResponseEntity<String>("Erreur", new HttpHeaders(), HttpStatus.NOT_FOUND);
         }
 
 
-        return new ResponseEntity("OK", new HttpHeaders(), HttpStatus.OK);
+        return new ResponseEntity<String>("OK", new HttpHeaders(), HttpStatus.OK);
     }
